move routes to views folder and remove SecondPage view
