--- src/views/SecondPage/SecondPage.style.tsx
@@ -1,10 +1,0 @@
-import { styled } from "@mui/material/styles";
-
-export const StyledPageWrapper = styled("div")`
-  margin-top: 80px;
-  display: flex;
-  gap: 16px;
-  flex-direction: column;
-  align-items: center;
-  min-height: 100vh;
-`;
