--- src/views/pages.ts
@@ -1,7 +1,0 @@
-export enum Pages {
-  Home = "/",
-  Secondary = "/second",
-  Projects = "/projects",
-  Owl_components = "/owl",
-  Owl_componentsInput = "/owlinput",
-}
