+++ src/views/routes.ts
